requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
PyJWT>=2.8,<3.0
bcrypt>=4.1,<5.0
waitress>=2.1,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<0.1
pydantic>=2.6,<3.0
